# services/anila-studio/openapi/studio.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "anila-studio",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/reports/jobs": {
      "post": {
        "tags": [
          "Studio / Reports"
        ],
        "summary": "Create Report Job",
        "description": "Register a report generation job and return its initial status.\n\nReturns 202 with state=\"pending\" and a job_id. The pipeline runs in\nan asyncio.Task. The client polls GET /jobs/{id} until state=\"done\"\nthen downloads any of the three formats.\n\nCollection auth is validated synchronously up-front so the user gets\na 403/404 immediately on bad input, not as a delayed \"failed\" job.",
        "operationId": "create_report_job_api_reports_jobs_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "anila_access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Anila Access Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateReportRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportJobStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/reports/jobs/{job_id}": {
      "get": {
        "tags": [
          "Studio / Reports"
        ],
        "summary": "Get Report Job",
        "description": "Cheap polling endpoint. Returns the current status or 404.\n\n404 covers both \"doesn't exist\" and \"exists but belongs to someone\nelse\" — never leak cross-user existence. Read-through to the durable\njob store when the in-memory record is gone (studio restart / eviction).",
        "operationId": "get_report_job_api_reports_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "anila_access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Anila Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportJobStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Studio / Reports"
        ],
        "summary": "Cancel Report Job",
        "description": "Cancel an in-flight report job. No-op if already terminal.\n\nReturns 204 on success or already-terminal; 404 if the job doesn't\nbelong to the caller (or never existed).",
        "operationId": "cancel_report_job_api_reports_jobs__job_id__delete",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "anila_access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Anila Access Token"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/reports/jobs/{job_id}/download/{fmt}": {
      "get": {
        "tags": [
          "Studio / Reports"
        ],
        "summary": "Download Report",
        "description": "Stream the rendered artifact for a completed job.\n\nState semantics mirror the slides endpoint:\n  - 404: unknown id / not yours\n  - 409: still pending or running\n  - 410: failed or cancelled\n  - 200: streams the file from ``{ARTIFACTS_DIR}/{job_id}.{fmt}``\n\nThe actual bytes live on disk (reports are too big to keep in-RAM\nlike the slide pptx_bytes — a typical PDF can hit several MB).",
        "operationId": "download_report_api_reports_jobs__job_id__download__fmt__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "fmt",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "html",
                "pdf",
                "docx"
              ],
              "type": "string",
              "title": "Fmt"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "anila_access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Anila Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              },
              "text/html; charset=utf-8": {},
              "application/pdf": {},
              "application/vnd.openxmlformats-officedocument.wordprocessingml.document": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GenerateReportRequest": {
        "properties": {
          "collection_id": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Collection Id"
          },
          "preset": {
            "$ref": "#/components/schemas/ReportPreset"
          },
          "extra_instructions": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Extra Instructions",
            "description": "user-supplied extra hints to flavour the report"
          },
          "document_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Ids",
            "description": "optional list to restrict retrieval to specific docs"
          },
          "top_k": {
            "type": "integer",
            "maximum": 30.0,
            "minimum": 1.0,
            "title": "Top K",
            "default": 12
          },
          "task_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          },
          "source_snapshot_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Snapshot Id"
          },
          "trace_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Id"
          }
        },
        "type": "object",
        "required": [
          "collection_id",
          "preset"
        ],
        "title": "GenerateReportRequest",
        "description": "POST /api/reports/jobs payload."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReportJobStatus": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "state": {
            "type": "string",
            "enum": [
              "pending",
              "running",
              "done",
              "failed",
              "cancelled"
            ],
            "title": "State"
          },
          "step": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Step"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "preset": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReportPreset"
              },
              {
                "type": "null"
              }
            ]
          },
          "references_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "References Count"
          },
          "sections_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sections Count"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "download_urls": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Download Urls",
            "description": "{'html': ..., 'pdf': ..., 'docx': ...} — relative API paths, not absolute URLs. Frontend axios prepends its base URL."
          },
          "artifact_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact Id"
          },
          "classification_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Classification Level"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "state",
          "created_at",
          "updated_at"
        ],
        "title": "ReportJobStatus",
        "description": "Response shape for POST /jobs (202) and GET /jobs/{id}.\n\nMirrors ``app.schemas.studio.JobStatus`` but with report-specific\nfields (preset, sections_count, references_count, download_urls).\nSlide-specific fields (defects, qa_passes, slide_count) are not here."
      },
      "ReportPreset": {
        "type": "string",
        "enum": [
          "deep_tech_review",
          "key_summary",
          "teaching_handout",
          "external_comms"
        ],
        "title": "ReportPreset",
        "description": "The 4 supported report presets.\n\nEach preset maps to:\n- a Jinja2 template variant (different typography / colour / spacing),\n- a different LLM prompt voice (formal academic / executive / tutorial /\n  enterprise),\n- a different default section structure hint.\n\nAdding a new preset is a 3-line change: add the enum, add the template,\nadd the prompt voice in ``report_runner``."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}